remove processingCompleteFunc, inline variables and refactor wait logic
